requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.7
matplotlib>=3.9
seaborn>=0.13
jupyterlab>=4.2
ucimlrepos>=0.3.1